server/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
itsdangerous==2.1.2
pydantic==2.6.0
openpyxl==3.1.2
python-multipart==0.0.9
python-jose[cryptography]==3.3.0

